refactor: Cart 삭제 시, 다른 회원의 장바구니 지우지 못하도록 변경

+++ src/main/java/cart/service/CartService.java
@@ -1,19 +1,18 @@
 package cart.service;
 
 import cart.controller.dto.CartResponse;
-import cart.controller.dto.ItemResponse;
 import cart.dao.CartDao;
 import cart.dao.ItemDao;
 import cart.dao.UserDao;
 import cart.domain.Cart;
 import cart.domain.Item;
 import cart.domain.User;
+import cart.exception.UnauthorizedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,6 +20,7 @@ public class CartService {
 
     public static final String NOT_EXIST_USER = "존재하지 않는 회원입니다.";
     public static final String NOT_EXIST_ITEM = "존재하지 않는 회원입니다.";
+    public static final String NOT_EXIST_CART = "존재하지 않는 장바구니 내역입니다.";
 
     private final CartDao cartDao;
     private final UserDao userDao;
@@ -43,14 +43,19 @@ public class CartService {
     @Transactional(readOnly = true)
     public List<CartResponse> loadItemInsideCart(final String email) {
         User user = userDao.findBy(email).orElseThrow(() -> new IllegalArgumentException(NOT_EXIST_USER));
-        List<Cart> carts = cartDao.findBy(user.getId());
+        List<Cart> carts = cartDao.findByUserId(user.getId());
         return carts.stream()
                 .map(CartResponse::from)
                 .collect(Collectors.toList());
     }
 
     @Transactional
-    public void deleteCart(final Long cartId) {
+    public void deleteCart(final String email, final Long cartId) {
+        User user = userDao.findBy(email).orElseThrow(() -> new IllegalArgumentException(NOT_EXIST_USER));
+        Cart cart = cartDao.findById(cartId).orElseThrow(() -> new IllegalArgumentException(NOT_EXIST_CART));
+        if (!Objects.equals(user.getId(), cart.getUser().getId())) {
+            throw new UnauthorizedException("권한이 없는 사용자 입니다.");
+        }
         cartDao.delete(cartId);
     }
 }
